Reworks to the API.
