--- src/__init__.py
@@ -1,6 +1,0 @@
-def main():
-    print("Hello World")
-
-
-if __name__ == "__main__":
-    main()
